refactor: Remove usage of query-builder in clients and comments middlewares
