+++ src/middleware/comments.ts
@@ -1,9 +1,8 @@
 import express from 'express'
-import queryBuilder, { Paginate, CriteriaPayload } from '../utils/query-builder'
 import CommentModel, { Comment } from '../db/models/comments'
 import { saveComment } from '../core/comments'
-import { setParams } from '../utils/paginate'
-import type { ParamsPayload } from '../utils/paginate'
+import { setParams, paginate, setQuery } from '../utils/paginate'
+import type { ParamsPayload, Paginate, CriteriaPayload } from '../utils/paginate'
 
 type CreatePayload = {
   comment: string
@@ -11,7 +10,6 @@ type CreatePayload = {
   batch?: string
 }
 
-const { paginate, setQuery } = queryBuilder('mongo')
 const index = async (
   req: express.Request<{ projectId: string }, {}, {}, CriteriaPayload>,
   res: express.Response<Paginate<Comment>>,
